Moved package.json reader file
