+++ environments/package-json.js
@@ -1,0 +1,3 @@
+
+// eslint-disable-next-line global-require
+export const packageJson = require('../package.json');
