--- environments/variables/package-json.js
@@ -1,3 +1,0 @@
-
-// eslint-disable-next-line global-require
-export const packageJson = require('../../package.json');
