+++ environments/variables/test.js
@@ -1,4 +1,4 @@
-import { packageJson } from './package-json';
+import { packageJson } from '../package-json';
 
 export const env = {
   ENVIRONMENT: 'test',
